Fix install script for XDG config files

+++ dotify.sh
@@ -46,8 +46,8 @@ function link {
 	done
 
   # Install config folders
-  ln -s ~/.dotfiles/fish.config ~/.config/fish
-  ln -s ~/.dotfiles/nvim.config ~/.config/nvim
+  [[ ! -L ~/.config/fish ]] && ln -s ~/.dotfiles/fish.config ~/.config/fish
+  [[ ! -L ~/.config/nvim ]] && ln -s ~/.dotfiles/nvim.config ~/.config/nvim
 
 }
